requirements.txt
requests
python-dotenv
playwright
reportlab
pdfplumber
